FEAT: Include a Product entity with pictures for sampling anonymous endpoints and downloading (#61)
## Description
Include a Product entity with pictures for sampling anonymous endpoints
and upload/download of files.

## Related Issue
Closes #52 

## How Has This Been Tested?
Template flag `--filesSupport` has been changed by
`--excludeFilesSupport`, which will default to true and will include the
implementation of the Products and Files by default and allows the user
to opt-out.
All the files and code related to Products and Files should be excluded
when passing the `--excludeFilesSupport` flag during the creation of a
new solution.

## Screenshots (if appropriate):

## Types of changes
- [ ] Bug fix (non-breaking change which fixes an issue)
- [X] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to change)

## Checklist:
<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->
- [X] My code follows the code style of this project.
- [X] My change requires a change to the documentation.
- [x] I have updated the documentation accordingly.
- [X] I have read the **CONTRIBUTING** document.
- [X] I have added tests to cover my changes.
- [X] All new and existing tests passed.